feat: add sd icon & update UI

+++ lib/widgets/spacing/sd_horizontal_spacing.dart
@@ -4,7 +4,7 @@ import 'package:system_design_flutter/resources/sd_spacing_constants.dart';
 class SdVerticalSpacing extends StatelessWidget {
   const SdVerticalSpacing({super.key, this.value, this.xRatio = 1});
   final double? value;
-  final int xRatio;
+  final double xRatio;
 
   @override
   Widget build(BuildContext context) {
